requirements.txt
# Python dependencies for BotOrNot

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0

# Deep Learning / NLP
torch>=1.10.0
transformers>=4.20.0
sentence-transformers>=2.2.0

# Data Processing
ijson>=3.1.4
tqdm>=4.62.0
python-dateutil>=2.8.2

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Model Persistence
joblib>=1.1.0

# File I/O
pyarrow>=9.0.0  # For parquet files

# Jupyter (optional, for notebook)
jupyter>=1.0.0
ipywidgets>=7.6.0
